Fix dependencies for html5lib

+++ debian/upgrade_2.3_to_2.5.1.sh
@@ -76,7 +76,7 @@ PG_OPENDATA_DATASTORE_BACKUP_NAME="backup_opendata_datastore_$(date '+%Y%m%d').d
 
 sudo -u postgres pg_dump opendata > $PG_OPENDATA_BACKUP_NAME
 sudo -u postgres pg_dump opendata_datastore > $PG_OPENDATA_DATASTORE_BACKUP_NAME
-paster --plugin=ckan db dump --config=/var/www/ckan/config/production.ini opendata_`date '+%Y%m%d_%H%M%S'`.pg_dump
+#paster --plugin=ckan db dump --config=/var/www/ckan/config/production.ini opendata_`date '+%Y%m%d_%H%M%S'`.pg_dump
 
 cp -r $CKAN_CONFIG_DIR $CURRENT_BACKUP_DIR
 cp -r $VIRTUALENV_DIR $CURRENT_BACKUP_DIR
@@ -113,7 +113,11 @@ pip install -e "git+$CKAN_REPO_TAG#egg=ckan"
 pip install -r "$VIRTUALENV_DIR/src/ckan/requirements.txt"
 pip install --upgrade -r "$VIRTUALENV_DIR/src/ckan/requirements.txt"
 
-pip uninstall html5lib
+if `pip show html5lib`
+then
+    pip uninstall html5lib
+fi
+
 pip install html5lib==0.9999999
 
 # make sure you’re using the virtualenv’s copies of commands like paster rather than any system-wide installed copies
